+++ src/note_label.rs
@@ -5,15 +5,15 @@ use serde::{Serialize, Deserialize};
 #[derive(Debug, FromRow, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct NoteLabel{
-    pub id: i64,
-    pub note_id: i64,
-    pub label_id: i64,
+    pub id: i32,
+    pub note_id: i32,
+    pub label_id: i32,
 }
 
 #[derive(Debug, FromRow, Serialize, Deserialize)]
 pub struct NewNoteLabel{
-    pub note_id: i64,
-    pub label_id: i64,
+    pub note_id: i32,
+    pub label_id: i32,
 }
 
 impl NoteLabel{
@@ -24,14 +24,14 @@ impl NoteLabel{
         Ok(notes_labels)
     }
 
-    pub async fn get(pool: web::Data<SqlitePool>, id: i64) -> Result<NoteLabel, Error>{
+    pub async fn get(pool: web::Data<SqlitePool>, id: i32) -> Result<NoteLabel, Error>{
         let note_label = query_as!(NoteLabel, r#"SELECT id, note_id, label_id FROM notes_labels WHERE id=$1"#, id)
             .fetch_one(pool.get_ref())
             .await?;
         Ok(note_label)
     }
 
-    pub async fn new(pool: web::Data<SqlitePool>, note_id: i64, label_id: i64) -> Result<NoteLabel, Error>{
+    pub async fn new(pool: web::Data<SqlitePool>, note_id: i32, label_id: i32) -> Result<NoteLabel, Error>{
         let id = query("INSERT INTO notes_labels (note_id, label_id) VALUES (?, ?);")
             .bind(note_id)
             .bind(label_id)
@@ -50,7 +50,7 @@ impl NoteLabel{
         Self::get(pool, note_label.id).await
     }
 
-    pub async fn delete(pool: web::Data<SqlitePool>, id: i64) -> Result<String, Error>{
+    pub async fn delete(pool: web::Data<SqlitePool>, id: i32) -> Result<String, Error>{
         query("DELETE FROM notes_labels WHERE id = ?;")
             .bind(id)
             .execute(pool.get_ref())
